Refactor tracking observers
