+++ Omikron/Factfinder/etc/frontend/events.xml
@@ -17,9 +17,9 @@
         <observer name="ff_export_authentication" instance="Omikron\Factfinder\Observer\ExportAuthentication" />
     </event>
     <event name="checkout_cart_add_product_complete">
-        <observer name="ff_add_to_cart" instance="Omikron\Factfinder\Observer\TrackingAddToCart" />
+        <observer name="ff_add_to_cart" instance="Omikron\Factfinder\Observer\Tracking\AddToCart" />
     </event>
     <event name="checkout_submit_all_after">
-        <observer name="ff_checkout" instance="Omikron\Factfinder\Observer\TrackingCheckout" />
+        <observer name="ff_checkout" instance="Omikron\Factfinder\Observer\Tracking\Checkout" />
     </event>
 </config>
